requirements.txt
pandas==1.0.3
scikit-learn==0.22.1
joblib==0.14.1
numpy==1.18.4
Flask==1.1.2
scipy==1.4.1
opencv-python==4.2.0.34
jupyter==1.0.0
Pillow==7.1.2
ibm-watson==4.4.0
azure.cognitiveservices.vision.face==0.4.0
python-dotenv==0.13.0
matplotlib==3.2.1